Moved inventory class to its own module
