+++ main.py
@@ -2,47 +2,9 @@
 
 import os
 import csv
+from inventory import Inventory
 from pprint import pprint
 DELIMITER = "\t"#assume tab-separated file, because that's my use case
-
-class Inventory():
-    """this is an abstraction of an inventory file
-    It contains a csv reader at Inventory.reader"""
-    def __init__(self, file):
-        self.f = open(file)
-        print("Loading file", file)
-        self.reader = csv.DictReader(self.f, delimiter=DELIMITER)
-        print("Found headers in inventory file:")
-        print(self.reader.fieldnames)
-        self.items = list(self.reader)
-        print("found", len(self.items), "items")
-        #self.close()#still trying to decide if necessary
-    def getvalue(self):
-        total = 0 
-        for each in self.items:
-            stock = float(each["STOCKONHAND"])
-            if stock <= 0: #Don't count items that aren't in stock
-                continue
-            cost = float(each["UNITCOST"])
-            total += stock * cost
-        return total        
-    def __len__(self):
-        return len(self.items)
-    def getDialect(self):
-        sniffer = csv.Sniffer()
-        sniffer.sniff(self)
-    def close(self):
-        self.f.close()
-    def findrecord(self, partnumber):
-        """takes a part number, returns the inventory entry for that item"""  
-        for row in self.items:
-            if row["PARTNUMBER"] == partnumber:
-                return row
-            if row["ALTPARTNUMBER"] == partnumber:
-                return row
-        #assume that if execution reached this point, nothing was found.
-        print("Could not find record,", partnumber)
-        return None
 
 def getcounts(filepath = ""):
     """This will read a list of barcodes, and flatten to a list of
